refactor: scope feature-dependent code to its features
The proxy, recording and remote code was defined unconditionally while every
caller was already behind a `#[cfg(feature = ...)]`, so most of it was dead in
all but a handful of feature combinations. Push the gates down to the
definitions, and group them so a feature is switched on in as few places as
possible:

- move the `StaticMockDefinition` YAML representation out of `common::data`
  into `common::static_mock`, gated once on `record`
- split `Recording` and `RecordingRuleBuilder` out of `api::proxy` into
  `api::record`, so both modules are gated once at their declaration
- split `server::state` into `state::proxy` and `state::record`, with the rule
  and recording maps held in private sub-structs of the state
- collect the feature-dependent methods of `MockServer`, `HttpMockServerBuilder`
  and the request handler into one `impl` block per feature

`MockServerAdapter` keeps a `#[cfg]` per method, since adapters are used as
trait objects and cannot be split up.
