+++ src/api/mock.rs
@@ -689,11 +689,14 @@ impl<'a> MockExt<'a> for Mock<'a> {
     }
 }
 
+/// A set of mocks created from a recording.
+#[cfg(feature = "record")]
 pub struct MockSet<'a> {
     pub ids: Vec<usize>,
     pub(crate) server: &'a MockServer,
 }
 
+#[cfg(feature = "record")]
 impl<'a> MockSet<'a> {
     pub fn delete(&mut self) {
         self.delete_async().join();
