+++ src/api/mod.rs
@@ -5,7 +5,7 @@ pub use mock::{Mock, MockExt};
 #[cfg(feature = "proxy")]
 pub use proxy::{ForwardingRule, ForwardingRuleBuilder, ProxyRule, ProxyRuleBuilder};
 #[cfg(feature = "record")]
-pub use proxy::{Recording, RecordingRuleBuilder};
+pub use record::{Recording, RecordingRuleBuilder};
 pub use server::MockServer;
 pub use spec::{Then, When};
 
@@ -14,7 +14,10 @@ use crate::common;
 mod adapter;
 mod mock;
 mod output;
+#[cfg(feature = "proxy")]
 mod proxy;
+#[cfg(feature = "record")]
+mod record;
 mod server;
 pub mod spec;
