+++ src/common/mod.rs
@@ -2,5 +2,8 @@ pub(crate) mod data;
 pub(crate) mod runtime;
 pub mod util;
 
+#[cfg(feature = "record")]
+pub(crate) mod static_mock;
+
 #[cfg(any(feature = "remote", feature = "proxy"))]
 pub mod http;
